+++ tests/apps/flask_app.py
@@ -1,10 +1,17 @@
-from flask import Flask, jsonify as J, Response, request
-from flask.views import MethodView
 import marshmallow as ma
+from flask import Flask, jsonify as J, Response, request, __version__ as flask_version
+from flask.views import MethodView
 
 from webargs import fields
-from webargs.flaskparser import parser, use_args, use_kwargs
 from webargs.core import json
+from webargs.flaskparser import (
+    parser,
+    use_args,
+    use_kwargs,
+)
+
+FLASK_MAJOR_VERSION = int(flask_version.split(".")[0])
+FLASK_SUPPORTS_ASYNC = FLASK_MAJOR_VERSION >= 2
 
 
 class TestAppConfig:
@@ -125,6 +132,14 @@ def echo_headers_raising(args):
     return J(args)
 
 
+if FLASK_SUPPORTS_ASYNC:
+
+    @app.route("/echo_headers_raising_async")
+    @use_args(HelloSchema(), location="headers", unknown=None)
+    async def echo_headers_raising_async(args):
+        return J(args)
+
+
 @app.route("/echo_cookie")
 def echo_cookie():
     return J(parser.parse(hello_args, request, location="cookies"))
@@ -144,10 +159,28 @@ def echo_view_arg(view_arg):
     return J(parser.parse({"view_arg": fields.Int()}, location="view_args"))
 
 
+if FLASK_SUPPORTS_ASYNC:
+
+    @app.route("/echo_view_arg_async/<view_arg>")
+    async def echo_view_arg_async(view_arg):
+        parsed_view_arg = await parser.async_parse(
+            {"view_arg": fields.Int()}, location="view_args"
+        )
+        return J(parsed_view_arg)
+
+
 @app.route("/echo_view_arg_use_args/<view_arg>")
 @use_args({"view_arg": fields.Int()}, location="view_args")
 def echo_view_arg_with_use_args(args, **kwargs):
     return J(args)
+
+
+if FLASK_SUPPORTS_ASYNC:
+
+    @app.route("/echo_view_arg_use_args_async/<view_arg>")
+    @use_args({"view_arg": fields.Int()}, location="view_args")
+    async def echo_view_arg_with_use_args_async(args, **kwargs):
+        return J(args)
 
 
 @app.route("/echo_nested", methods=["POST"])
@@ -172,6 +205,18 @@ def echo_nested_many_with_data_key():
     return J(parser.parse(args))
 
 
+if FLASK_SUPPORTS_ASYNC:
+
+    @app.route("/echo_nested_many_data_key_async", methods=["POST"])
+    async def echo_nested_many_with_data_key_async():
+        args = {
+            "x_field": fields.Nested(
+                {"id": fields.Int()}, many=True, data_key="X-Field"
+            )
+        }
+        return J(await parser.async_parse(args))
+
+
 class EchoMethodViewUseArgs(MethodView):
     @use_args({"val": fields.Int()})
     def post(self, args):
@@ -182,6 +227,19 @@ app.add_url_rule(
     "/echo_method_view_use_args",
     view_func=EchoMethodViewUseArgs.as_view("echo_method_view_use_args"),
 )
+
+
+if FLASK_SUPPORTS_ASYNC:
+
+    class EchoMethodViewUseArgsAsync(MethodView):
+        @use_args({"val": fields.Int()})
+        async def post(self, args):
+            return J(args)
+
+    app.add_url_rule(
+        "/echo_method_view_use_args_async",
+        view_func=EchoMethodViewUseArgsAsync.as_view("echo_method_view_use_args_async"),
+    )
 
 
 class EchoMethodViewUseKwargs(MethodView):
@@ -195,12 +253,35 @@ app.add_url_rule(
     view_func=EchoMethodViewUseKwargs.as_view("echo_method_view_use_kwargs"),
 )
 
+if FLASK_SUPPORTS_ASYNC:
+
+    class EchoMethodViewUseKwargsAsync(MethodView):
+        @use_kwargs({"val": fields.Int()})
+        async def post(self, val):
+            return J({"val": val})
+
+    app.add_url_rule(
+        "/echo_method_view_use_kwargs_async",
+        view_func=EchoMethodViewUseKwargsAsync.as_view(
+            "echo_method_view_use_kwargs_async"
+        ),
+    )
+
 
 @app.route("/echo_use_kwargs_missing", methods=["post"])
 @use_kwargs({"username": fields.Str(required=True), "password": fields.Str()})
 def echo_use_kwargs_missing(username, **kwargs):
     assert "password" not in kwargs
     return J({"username": username})
+
+
+if FLASK_SUPPORTS_ASYNC:
+
+    @app.route("/echo_use_kwargs_missing_async", methods=["post"])
+    @use_kwargs({"username": fields.Str(required=True), "password": fields.Str()})
+    async def echo_use_kwargs_missing_async(username, **kwargs):
+        assert "password" not in kwargs
+        return J({"username": username})
 
 
 # Return validation errors as JSON