+++ tests/test_djangoparser.py
@@ -1,5 +1,6 @@
 import pytest
 from tests.apps.django_app.base.wsgi import application
+from tests.apps.django_app import DJANGO_SUPPORTS_ASYNC
 
 from webargs.testing import CommonTestCase
 
@@ -27,3 +28,15 @@ class TestDjangoParser(CommonTestCase):
     def test_use_args_in_class_based_view_with_path_param(self, testapp):
         res = testapp.get("/echo_use_args_with_path_param_cbv/42?name=Fred")
         assert res.json == {"name": "Fred"}
+
+    @pytest.mark.skipif(
+        not DJANGO_SUPPORTS_ASYNC, reason="requires a django version with async support"
+    )
+    def test_parse_querystring_args_async(self, testapp):
+        assert testapp.get("/async_echo?name=Fred").json == {"name": "Fred"}
+
+    @pytest.mark.skipif(
+        not DJANGO_SUPPORTS_ASYNC, reason="requires a django version with async support"
+    )
+    def test_async_use_args_decorator(self, testapp):
+        assert testapp.get("/async_echo_use_args?name=Fred").json == {"name": "Fred"}